# Conflicts:
#	src/main/java/Venue_Event_Manager/repository/jdbc/PgUserRepository.java

+++ src/main/java/Venue_Event_Manager/repository/UserRepository.java
@@ -22,7 +22,7 @@ public interface UserRepository {
 
     List<User> findAllByAccountStatus(Connection conn, AccountStatus accountStatus);
 
-    Optional<Integer> getAverageReview(Connection conn, long userId);
+    Optional<Double> getAverageReview(Connection conn, long userId);
 
     Optional<String> getPasswordById(Connection conn, long userId);
